Change logging saved folder
For production use, log should be on a folder named 'events'
on SDCard root.
If a file called 'info' is on events folder, log level is info
If a file called 'debug' is on events folder, log level is debug
If no file is present, there is no log file.